IOS-00064 Plans IOS-00064B Refactor ActivitiesFeature
