--- WorkoutMirrorLive/FeaturesNew/ActivitiesTab/Activities/Feature/ActivitiesFeature+Destination.swift
@@ -1,22 +1,0 @@
-//
-//  ActivitiesFeature+Destination.swift
-//  MyFitnessJournal
-//
-//  Created by Sebastian Sciuba on 15/12/2025.
-//
-
-import ComposableArchitecture
-
-/// Implementation of `ActivitiesFeature` destination
-extension ActivitiesFeature {
-    
-    @Reducer
-    enum Destination {
-        
-        ///
-        case zoneInfo(HeartRateZoneInfoFeature)
-        
-        ///
-        case activityDetails(ActivityDetailsFeature)
-    }
-}
